+++ fda-model/alternatives/Alternative.cs
@@ -123,9 +123,9 @@ namespace alternatives
             {
                 masterseed = randomProvider.Seed;
             }
-            int progressChunks = 1;
-            int _completedIterations = 0;
-            int _ExpectedIterations = convergenceCriteria.MaxIterations;
+            Int64 progressChunks = 1;
+            Int64 _completedIterations = 0;
+            Int64 _ExpectedIterations = convergenceCriteria.MaxIterations;
             if (_ExpectedIterations > 100)
             {
                 progressChunks = _ExpectedIterations / 100;
@@ -136,7 +136,7 @@ namespace alternatives
             {
                 seeds[i] = masterSeedList.Next();
             }
-            int iterations = convergenceCriteria.MinIterations;
+            Int64 iterations = convergenceCriteria.MinIterations;
 
             while (!aaeqResult.ConsequenceHistogram.IsConverged)
             {
